+++ src/components/Top.js
@@ -1,23 +1,6 @@
 import React, { useState } from "react";
 import { makeStyles } from "@material-ui/core/styles";
-import AppBar from "@material-ui/core/AppBar";
-import Toolbar from "@material-ui/core/Toolbar";
-import Typography from "@material-ui/core/Typography";
-import Button from "@material-ui/core/Button";
-import IconButton from "@material-ui/core/IconButton";
-import MenuIcon from "@material-ui/icons/Menu";
-import AccountBox from "@material-ui/icons/AccountBox";
 import { useHistory, useLocation } from "react-router-dom";
-import Drawer from "@material-ui/core/Drawer";
-import List from "@material-ui/core/List";
-import ListItem from "@material-ui/core/ListItem";
-import ListItemIcon from "@material-ui/core/ListItemIcon";
-import ListItemText from "@material-ui/core/ListItemText";
-import Divider from "@material-ui/core/Divider";
-import AddBoxIcon from "@material-ui/icons/AddBox";
-import CreateIcon from "@material-ui/icons/Create";
-import ChatIcon from "@material-ui/icons/Chat";
-import NotificationsIcon from "@material-ui/icons/Notifications";
 
 //Chart
 import {
@@ -95,98 +78,9 @@ function Top() {
   const history = useHistory();
   const locate = useLocation();
   const [value, setValue] = React.useState(locate.pathname.slice(1));
-  const [state, setState] = useState({ left: false });
-  const toggleDrawer = (side, open) => event => {
-    if (
-      event.type === "keydown" &&
-      (event.key === "Tab" || event.key === "Shift")
-    ) {
-      return;
-    }
 
-    setState({ ...state, [side]: open });
-  };
-  const menuClick = value => event => {
-    return alert(value);
-  };
-
-  const sideList = side => (
-    <div
-      className={classes.list}
-      role="presentation"
-      onClick={toggleDrawer(side, false)}
-      onKeyDown={toggleDrawer(side, false)}
-    >
-      <List>
-        <ListItem>
-          <ListItemText primary="目標" />
-        </ListItem>
-        <ListItem button onClick={menuClick("targetregist")}>
-          <ListItemIcon>
-            <AddBoxIcon />
-          </ListItemIcon>
-          <ListItemText primary="目標追加" />
-        </ListItem>
-        <ListItem button key="regist" onClick={menuClick()}>
-          <ListItemIcon>
-            <CreateIcon />
-          </ListItemIcon>
-          <ListItemText primary="実績追加" />
-        </ListItem>
-      </List>
-      <Divider />
-      <List>
-        <ListItem>
-          <ListItemText primary="SNS" />
-        </ListItem>
-        <ListItem button key="chat" onClick={menuClick}>
-          <ListItemIcon>
-            <ChatIcon />
-          </ListItemIcon>
-          <ListItemText primary="Chat" />
-        </ListItem>
-      </List>
-      <Divider />
-      <List>
-        <ListItem>
-          <ListItemText primary="通知" />
-        </ListItem>
-        <ListItem button key="notify" onClick={menuClick}>
-          <ListItemIcon>
-            <NotificationsIcon />
-          </ListItemIcon>
-          <ListItemText primary="通知" />
-        </ListItem>
-      </List>
-    </div>
-  );
   return (
     <div className={classes.root}>
-      <header className={classes.header}>
-        <Drawer open={state.left} onClose={toggleDrawer("left", false)}>
-          {sideList("left")}
-        </Drawer>
-        <AppBar>
-          <Toolbar>
-            <IconButton
-              edge="start"
-              className={classes.menuButton}
-              color="inherit"
-              aria-label="menu"
-              onClick={toggleDrawer("left", true)}
-            >
-              <MenuIcon />
-            </IconButton>
-            <Typography variant="h6" className={classes.title}>
-              Top
-            </Typography>
-            <Button color="inherit" onClick={() => history.replace("/logout")}>
-              LogOut
-            </Button>
-            <AccountBox></AccountBox>
-          </Toolbar>
-        </AppBar>
-      </header>
       <div className={classes.contents}>
         <h1>TOP PAGE!!!!!!!!!!</h1>
         <h2>取得者年次分布</h2>
